added more tests for Game

+++ ConsoleRoyale/ConsoleRoyale/RNG/IRNGProvider.cs
@@ -18,11 +18,11 @@
         double GetRandomDouble();
 
         /// <summary>
-        /// Get random double in a given interval
+        /// Get random floating-point number in a given interval
         /// </summary>
         /// <param name="start">The start of the interval</param>
         /// <param name="end">The end of the interval</param>
-        /// <returns>A random double number in the given interval</returns>
-        double GetRandomDoubleInInterval(int start, int end);
+        /// <returns>A random floating-point number in the given interval</returns>
+        decimal GetRandomDoubleInInterval(int start, int end);
     }
 }
